Change scanner module structure
