--- rust/nasl-interpreter/src/scanner/scan_interpreter.rs
@@ -1,3 +1,0 @@
-// SPDX-FileCopyrightText: 2024 Greenbone AG
-//
-// SPDX-License-Identifier: GPL-2.0-or-later WITH x11vnc-openssl-exception
